+++ src/paddlefx/translator.py
@@ -1,9 +1,5 @@
 from __future__ import annotations
 
-import dataclasses
-import dis
-import itertools
-import logging
 import operator
 import types
 
@@ -12,79 +8,9 @@ from typing import Any
 import paddle
 import paddle.nn
 
-from .graph_layer import GraphLayer
+from .bytecode_transformation import Instruction, create_instruction
+from .output_graph import OutputGraph
 from .proxy import Attribute, Proxy
-from .symbolic_trace import Tracer
-
-__all__ = ['OutputGraph', 'Instruction', 'InstructionTranslator', 'convert_instruction']
-
-# TODO: better code organization
-
-
-@dataclasses.dataclass
-class InstructionExnTabEntry:
-    start: Instruction
-    end: Instruction
-    target: Instruction
-    depth: int
-    lasti: bool
-
-    def __repr__(self):
-        return (
-            f"InstructionExnTabEntry(start={self.start.short_inst_repr()}, "
-            f"end={self.end.short_inst_repr()}, "
-            f"target={self.target.short_inst_repr()}, "
-            f"depth={self.depth}, lasti={self.lasti})"
-        )
-
-    def __eq__(self, o):
-        return (
-            self.start is o.start
-            and self.end is o.end
-            and self.target is o.target
-            and self.depth == o.depth
-            and self.lasti == o.lasti
-        )
-
-
-@dataclasses.dataclass
-class Instruction:
-    """A mutable version of dis.Instruction."""
-
-    opcode: int
-    opname: str
-    arg: int | None
-    argval: Any
-    offset: int | None = None
-    starts_line: int | None = None
-    is_jump_target: bool = False
-    # extra fields to make modification easier:
-    target: Instruction | None = None
-    exn_tab_entry: InstructionExnTabEntry | None = None
-
-    def __hash__(self):
-        return id(self)
-
-    def __eq__(self, other):
-        return id(self) == id(other)
-
-
-def convert_instruction(i: dis.Instruction):
-    return Instruction(
-        i.opcode,
-        i.opname,
-        i.arg,
-        i.argval,
-        i.offset,
-        i.starts_line,
-        i.is_jump_target,
-    )
-
-
-def create_instruction(name, *, arg=None, argval=None, target=None):
-    return Instruction(
-        opcode=dis.opmap[name], opname=name, arg=arg, argval=argval, target=target
-    )
 
 
 def _binary_constructor(op_name: str):
@@ -111,109 +37,6 @@ def _not_implemented(op_name):
         raise NotImplementedError()
 
     return _not_impl
-
-
-_unique_id_counter = itertools.count()
-
-
-def unique_id(name):
-    return f"{name}_{next(_unique_id_counter)}"
-
-
-class OutputGraph(Tracer):
-    def __init__(
-        self,
-        *,
-        f_globals: dict[str, Any],
-        code_options: dict,
-        compiler_fn: Any,
-    ):
-        super().__init__()
-
-        self.f_globals = f_globals
-        self.code_options = code_options
-        self.compiler_fn = compiler_fn
-
-        self.output_instructions: list[Instruction] = []
-
-    @property
-    def placeholders(self):
-        r = []
-        for node in self.graph.nodes:
-            if node.op == "placeholder":
-                r.append(node)
-                continue
-            break
-        return r
-
-    def install_global(self, name, value) -> None:
-        self.f_globals[name] = value
-
-    def update_co_names(self, name: str):
-        if name not in self.code_options["co_names"]:
-            self.code_options["co_names"] += (name,)
-
-    def add_output_instructions(self, prefix: list[Instruction]) -> None:
-        self.output_instructions.extend(prefix)
-
-    def call_user_compiler(self, gl):
-        compiled_fn = self.compiler_fn(gl)
-        return compiled_fn
-
-    def compile_and_call_fx_graph(self, tx, rv, root):
-        from .codegen import PyCodegen
-        from .eval_frame import disable
-
-        self.create_node("output", "output", tuple(x for x in rv), {})
-
-        gl = GraphLayer(root, self.graph)
-
-        compiled_fn = self.call_user_compiler(gl)
-        compiled_fn = disable(compiled_fn)
-
-        name = unique_id("__compiled_fn")
-
-        logging.debug(f"{name} - gl.src:\n{gl.src}")
-
-        logging.debug(f"{name}:")
-        [logging.debug(x) for x in list(dis.get_instructions(compiled_fn))]
-        logging.debug(f"")
-
-        logging.debug(f"{name}.fn:")
-        [logging.debug(x) for x in list(dis.get_instructions(compiled_fn.fn))]
-        logging.debug(f"")
-
-        self.install_global(name, compiled_fn)
-
-        cg = PyCodegen(tx)
-        cg.make_call_generated_code(name)
-        return cg.get_instructions()
-
-    def compile_subgraph(
-        self,
-        tx: InstructionTranslatorBase,
-    ):
-        stack_values = list(tx.stack)
-
-        if not (root := tx.frame.f_locals.get('self', None)):
-            root = paddle.nn.Layer()
-
-        instructions = []
-        instructions.extend(
-            self.compile_and_call_fx_graph(
-                tx,
-                list(reversed(stack_values)),
-                root,
-            )
-        )
-
-        # for outputs == 0
-        # instructions.append(create_instruction("POP_TOP"))
-
-        # for outputs != 0
-        # instructions.append(create_instruction("UNPACK_SEQUENCE", arg=1))
-
-        self.add_output_instructions(instructions)
 
 
 BINARY_MAPPER = {
@@ -299,9 +122,7 @@ class InstructionTranslatorBase:
                 is_custom_call = True
                 break
 
-        if fn is print:
-            self.push(None)
-        elif isinstance(fn, Attribute):
+        if isinstance(fn, Attribute):
             self.push(fn(*args, **kwargs))
         elif fn is isinstance:
             res = self.output.create_node('call_function', fn, args, kwargs)
@@ -506,7 +327,7 @@ class InstructionTranslator(InstructionTranslatorBase):
 
     def step(self, inst: Instruction):
         if not hasattr(self, inst.opname):
-            raise Exception(f"missing: {inst.opname}")
+            raise NotImplementedError(f"missing: {inst.opname}")
         getattr(self, inst.opname)(inst)
 
     def run(self):
